requirements.txt
Flask~=3.0.3
Flask-SQLAlchemy~=3.1.1
numpy~=1.26.4
weaviate-client~=4.4.4
PyPDF2~=3.0.1
nltk~=3.8.1
sentence-transformers~=2.3.1
